Controlled x-data sorting #6450
- Removed unordered toMap() call and replaced with more accurate and effective code
- Moved the only method in XDataResource to XDataContextResource and removed the former
