+++ modules/rest/src/main/java/com/enonic/xp/app/contentstudio/rest/resource/content/CmsResourceDynamicFeature.java
@@ -17,7 +17,6 @@ import com.enonic.xp.app.contentstudio.rest.resource.schema.content.ContentTypeC
 import com.enonic.xp.app.contentstudio.rest.resource.schema.content.ContentTypeResource;
 import com.enonic.xp.app.contentstudio.rest.resource.schema.content.FilterByContentResource;
 import com.enonic.xp.app.contentstudio.rest.resource.schema.xdata.XDataContextResource;
-import com.enonic.xp.app.contentstudio.rest.resource.schema.xdata.XDataResource;
 import com.enonic.xp.jaxrs.JaxRsComponent;
 
 
@@ -28,7 +27,7 @@ public final class CmsResourceDynamicFeature
     private final Set<Class<? extends JaxRsComponent>> supportedResources =
         Set.of( ContentResource.class, ContentImageResource.class, ContentIconResource.class, ContentMediaResource.class,
                 IssueResource.class, FilterByContentResource.class, ContentTypeResource.class, PageResource.class,
-                PageTemplateResource.class, FragmentResource.class, XDataResource.class, XDataContextResource.class,
+                PageTemplateResource.class, FragmentResource.class, XDataContextResource.class,
                 ArchiveResource.class, ContentTypeContextResource.class, MacroResource.class );
