+++ modules/rest/src/test/resources/com/enonic/xp/app/contentstudio/rest/resource/schema/xdata/get_content_x_data.json
@@ -3,34 +3,6 @@
         {
             "createdTime": "2013-01-01T12:00:00Z",
             "deletable": false,
-            "displayName": "text_area_2",
-            "editable": false,
-            "form": {
-                "formItems": [
-                    {
-                        "Input": {
-                            "config": {},
-                            "helpText": "Help text area",
-                            "immutable": false,
-                            "indexed": false,
-                            "inputType": "TextArea",
-                            "label": "Text Area",
-                            "maximizeUIInputWidth": true,
-                            "name": "text_area_2",
-                            "occurrences": {
-                                "maximum": 1,
-                                "minimum": 1
-                            }
-                        }
-                    }
-                ]
-            },
-            "isOptional": false,
-            "name": "myapplication:text_area_2"
-        },
-        {
-            "createdTime": "2013-01-01T12:00:00Z",
-            "deletable": false,
             "displayName": "input_text_1",
             "editable": false,
             "form": {
@@ -55,6 +27,34 @@
             },
             "isOptional": false,
             "name": "myapplication:input_text_1"
+        },
+        {
+            "createdTime": "2013-01-01T12:00:00Z",
+            "deletable": false,
+            "displayName": "text_area_2",
+            "editable": false,
+            "form": {
+                "formItems": [
+                    {
+                        "Input": {
+                            "config": {},
+                            "helpText": "Help text area",
+                            "immutable": false,
+                            "indexed": false,
+                            "inputType": "TextArea",
+                            "label": "Text Area",
+                            "maximizeUIInputWidth": true,
+                            "name": "text_area_2",
+                            "occurrences": {
+                                "maximum": 1,
+                                "minimum": 1
+                            }
+                        }
+                    }
+                ]
+            },
+            "isOptional": false,
+            "name": "myapplication:text_area_2"
         }
     ]
 }
